+++ controllers/productController.js
@@ -2,9 +2,9 @@ const db = require("../confic/db");
 const { product } = db;
 const Sequelize = require("sequelize");
 
-const getProduct = async (req, res) => {
+const getProducts = async (req, res) => {
   try {
-    let products = await product.findAll({ where: {} });
+    let products = await product.findAll();
     res.status(200).json({ products: products });
   } catch (error) {
     return res.status(400).json({
@@ -13,18 +13,19 @@ const getProduct = async (req, res) => {
   }
 };
 
-const addProduct = async (req, res) => {
+const addProducts = async (req, res) => {
   try {
-    const { productName, description, category, price } = req.body;
+    const { name, description, category, price, image } = req.body;
 
-    if (!productName && description && category && price) {
+    if (!name && description && category && price && image) {
       throw new Error("All fields are required");
     } else {
       const createProduct = await product.create({
-        productName: productName,
+        name: name,
         description: description,
         category: category,
         price: price,
+        image: image,
       });
       let products = await createProduct.save();
       return res.status(200).json({
@@ -41,15 +42,25 @@ const addProduct = async (req, res) => {
 
 const deleteProduct = async (req, res) => {
   try {
-    let products = await product.destroy({
+    const { name } = req.body;
+    let allReadyExistProduct = product.findAll({
       where: {
-        id: req.params.id,
+        name: product.name,
       },
     });
-    return res.status(200).json({
-      message: "Product delete successfully",
-      products: products,
-    });
+    if (allReadyExistProduct) {
+      let products = await product.destroy({
+        where: {
+          id: req.params.id,
+        },
+      });
+      return res.status(200).json({
+        message: "Product delete successfully",
+        products: products,
+      });
+    } else {
+      throw new Error("Product not found");
+    }
   } catch (error) {
     return res.status(400).json({
       message: error.message,
@@ -57,7 +68,7 @@ const deleteProduct = async (req, res) => {
   }
 };
 
-const getOneProduct = async (req, res) => {
+const getProduct = async (req, res) => {
   try {
     let products = await product.findOne({
       where: {
@@ -95,8 +106,8 @@ const searchProduct = async (req, res) => {
   try {
     let result = await product.findAll({
       where: {
-        productName: { [Sequelize.Op.regexp]: req.params.key }
-      }
+        productName: { [Sequelize.Op.regexp]: req.params.key },
+      },
     });
     return res.status(200).json({
       result: result,
@@ -109,10 +120,10 @@ const searchProduct = async (req, res) => {
 };
 
 module.exports = {
-  getProduct,
-  addProduct,
+  getProducts,
+  addProducts,
   deleteProduct,
-  getOneProduct,
+  getProduct,
   updateProduct,
   searchProduct,
 };
